+++ ct-app/tests/test_other_endpoints.py
@@ -1,3 +1,4 @@
+import asyncio
 import pytest
 import aiohttp
 from unittest.mock import patch
@@ -28,10 +29,15 @@ async def test_blacklist_rpch_nodes():
             "some_rpch_endpoint",
             "some_subgraph_url",
         )
+        node.started = True
 
-        result = await node.blacklist_rpch_nodes("some_api_endpoint")
+        asyncio.create_task(node.blacklist_rpch_nodes())
+        await asyncio.sleep(0.5)
 
-        assert result == ("rpch", ["1", "2", "3"])
+        node.started = False
+        await asyncio.sleep(0.5)
+
+        assert node.rpch_nodes == ["1", "2", "3"]
 
 
 @pytest.mark.asyncio
@@ -49,8 +55,11 @@ async def test_blacklist_rpch_nodes_exceptions():
             "some_rpch_endpoint",
             "some_subgraph_url",
         )
-        result = await node.blacklist_rpch_nodes("some_api_endpoint")
-        assert result == ("rpch", [])
+        asyncio.create_task(node.blacklist_rpch_nodes())
+        await asyncio.sleep(0.5)
+        node.started = False
+        await asyncio.sleep(0.5)
+        assert node.rpch_nodes is None
 
         # Simulate ValueError
         mock_get.side_effect = OSError("ValueError")
@@ -60,8 +69,11 @@ async def test_blacklist_rpch_nodes_exceptions():
             "some_rpch_endpoint",
             "some_subgraph_url",
         )
-        result = await node.blacklist_rpch_nodes("some_api_endpoint")
-        assert result == ("rpch", [])
+        asyncio.create_task(node.blacklist_rpch_nodes())
+        await asyncio.sleep(0.5)
+        node.started = False
+        await asyncio.sleep(0.5)
+        assert node.rpch_nodes is None
 
         # Simulate general exception
         mock_get.side_effect = Exception("Exception")
@@ -71,8 +83,11 @@ async def test_blacklist_rpch_nodes_exceptions():
             "some_rpch_endpoint",
             "some_subgraph_url",
         )
-        result = await node.blacklist_rpch_nodes("some_api_endpoint")
-        assert result == ("rpch", [])
+        asyncio.create_task(node.blacklist_rpch_nodes())
+        await asyncio.sleep(0.5)
+        node.started = False
+        await asyncio.sleep(0.5)
+        assert node.rpch_nodes is None
 
 
 # @pytest.mark.asyncio
